.tools/requirements.txt
anthropic>=0.25.0
networkx>=3.2
watchdog>=4.0
python-dotenv>=1.0.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.120.0
